# server/routes/swagger.yaml
openapi: 3.0.0
info:
  title: SnipHub API
  description: API documentation for SnipHub code snippet sharing platform
  version: 1.0.0
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
        email:
          type: string
        favorites:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    Snippet:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        code:
          type: string
        programmingLanguage:
          type: string
        description:
          type: string
        user:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/UserReference'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserReference:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
    Error:
      type: object
      properties:
        message:
          type: string
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Email or username
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /auth/me:
    get:
      summary: Get current user profile
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/favorites:
    get:
      summary: Get all favorites for current user
      tags:
        - Users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of favorite snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snippet'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/favorites/{snippetId}:
    post:
      summary: Add a snippet to user's favorites
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: snippetId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated favorites list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Already in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove a snippet from user's favorites
      tags:
        - Users
      security:
        - bearerAuth: []
      parameters:
        - name: snippetId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated favorites list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: Not in favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/{id}:
    get:
      summary: Get user profile by ID
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /users/{id}/snippets:
    get:
      summary: Get all snippets created by a user
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of user snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snippet'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /snippets:
    get:
      summary: Get all snippets
      tags:
        - Snippets
      parameters:
        - name: language
          in: query
          schema:
            type: string
          description: Filter by programming language
        - name: search
          in: query
          schema:
            type: string
          description: Search term for title, description or code
      responses:
        '200':
          description: List of snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Snippet'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new snippet
      tags:
        - Snippets
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - code
              properties:
                title:
                  type: string
                code:
                  type: string
                description:
                  type: string
                programmingLanguage:
                  type: string
                language:
                  type: string
                  description: Alternative field for programmingLanguage
      responses:
        '201':
          description: Snippet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /snippets/{id}:
    get:
      summary: Get a specific snippet by ID
      tags:
        - Snippets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Snippet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '404':
          description: Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a snippet
      tags:
        - Snippets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                code:
                  type: string
                description:
                  type: string
                programmingLanguage:
                  type: string
                language:
                  type: string
                  description: Alternative field for programmingLanguage
      responses:
        '200':
          description: Updated snippet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Snippet'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not the snippet owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a snippet
      tags:
        - Snippets
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Snippet deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (not the snippet owner)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Snippet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /snippets/languages/list:
    get:
      summary: Get list of all used programming languages
      tags:
        - Snippets
      responses:
        '200':
          description: List of programming languages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'